[12.09.24]:[4:2]: inflater with allocator and rebinding u8 -> u32 for inherited allocator

+++ src/csc/png/serializer/produce_chunk/deflater/deflater.lib.cxx
@@ -47,6 +47,6 @@ class deflater : private deflater_impl<Alloc> {
   }
 };
 
-using common_deflater = csc::deflater<std::allocator<uint32_t>>;
+using common_deflater = csc::deflater<std::allocator<uint8_t>>;
 
 } // namespace csc
